tests: change adaid type to smallint, to easy cross-db tests

+++ tests/schema.sql
@@ -3,7 +3,7 @@ CREATE DATABASE IF NOT EXISTS `phpstan_dba`;
 USE `phpstan_dba`;
 
 CREATE TABLE `ada` (
-    `adaid` int(11) NOT NULL,
+    `adaid` smallint NOT NULL,
     `gesperrt` tinyint(1) NOT NULL DEFAULT '0',
     `email` varchar(100) NOT NULL DEFAULT '',
     `freigabe1u1` smallint(1) NOT NULL
